+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 
 ## Improvements and additions
 
+- bind Ctrl+scrollwheel to scroll-left/right; change to move cursor by `options.scroll_incr` (like scroll-up/down)
 - split `Ctrl+V` (check-version) into `Ctrl+V` (show-version) and `z Ctrl+V` (require-version)
 - `q`uit scrubs cmdlog of the dropped sheet if it was a leaf/orphan (no other sheet has it as a source)
 - numeric binning for frequency/pivot table; `options.histogram_bins` to set number of bins (0 to choose a reasonable default)